feat(driver): QRE1113 sensor application code
Application code, line.c and line.h to detect the postion of the white
line using the qre1113 driver. The line is detected by setting a
threshold voltage value that corresponds to white or dark being
detected. When below the threshold the white line is detected. The line
position is gotten by filtering through the 4 sensors to set what
sensors detects a white line and making a direction decision based on the
sensors that detected and their position on the SUMO bot.