feat(driver): I2C driver for VL530X sensor
I2C driver for VL530X sensor using polling, since use case for the
sensor does not require interrupts. The driver includes i2c read and write functionality. The driver works for multi byte address and data applications, but 1-3 bytes will be the range utilized for the this project when interacting with the VL530X sensor. The test function for the driver includes reading and writing to and from writiable/readable addresses of VL530X sensor.