feat(drivers): URM09 Range Sensor
Created driver for the URM09 sensor to interface the I2C driver. The
driver makes use of a set measurement function to put the sensor in
either automatic or passive mode, in this application automatic mode
will be used. The driver uses the get distance function to collect the
range value in cm, from 2 registers in the sensor. An out of range value
for the sensor is defined as 65535.

Minor changes are also made to other driver files, to most clean
commented code and update function names (i2c).